+++ backend/connector-integration/src/connectors.rs
@@ -55,3 +55,6 @@ pub use self::braintree::Braintree;
 
 pub mod volt;
 pub use self::volt::Volt;
+
+pub mod bluecode;
+pub use self::bluecode::Bluecode;